refactor: refactor code for submit (#33)
